+++ lib/parsers/index.js
@@ -1,0 +1,5 @@
+require('./yaml');
+require('./json');
+require('./toml')
+require('./javascript');
+module.exports = require('./coffee');
