+++ lib/parsers/js-yaml.js
