+++ lib/parsers/msg.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function msg(lang, err) {
+  return 'gray-matter parser [' + lang + ']: ' + err;
+}
